Revert "BF-90 deisgn with real data and tested"
This reverts commit 2908b667b8059b351236435a631222dbfa049ca9.

+++ frontend/lib/models/user_profile.dart
@@ -5,9 +5,9 @@ class UserProfile {
 
   final String email;
 
-  final double height;
+  final Double height;
 
-  final double weight;
+  final Double weight;
 
   final String birthdate;
 
@@ -22,8 +22,8 @@ class UserProfile {
     return UserProfile(
         username: json['username'] as String,
         email: json['email'] as String,
-        height: json['height'],
-        weight: json['weight'],
+        height: json['height'] as Double,
+        weight: json['weight'] as Double,
         birthdate: json['birthdate'] as String);
   }
